test(export): pin the properties geom_color_key's dedup actually relies on
The follow-up promised when reviewing #2226 (the XXH3 BSL-1.0 -> MIT swap),
now that it has landed.

That PR's own argument was that the digest gates key equality for dedup
grouping and never reaches the output, so any sound 128-bit hash groups the
same meshes. Correct — and it is exactly why NO test in the suite could
corroborate the swap. The 17-model byte-identical corpus comparison was doing
all the verification; the 471 passing tests said nothing about it.

What that leaves unguarded is a FUTURE swap to something degenerate.
`geom_color_key` had no test at all: it appeared only at its definition and
three call sites. A replacement returning a constant, or ignoring the tail of
its input, would pass everything while silently merging distinct meshes into
one glTF primitive — geometry that vanishes or takes the wrong material, and
only on models containing near-identical parts.

Six fixture-free tests pin what the dedup depends on: one position bit,
colour alone, normals alone, indices alone, bit-identical meshes collapsing,
and the length-framing guard — the `u64` written before each run so two
different (positions, normals, indices) splits of the same bytes cannot
alias.

Each is justified by a mutation shown to survive before the test existed:

- `finish_128()` -> a constant: 10 failures now
- all four colour channels dropped: kills the colour test specifically
- one length frame removed: kills the aliasing test specifically

Note the first colour mutation attempted was too weak — zeroing only the red
channel, while the fixture's two colours differ in red AND green, so green
still distinguished them. Dropping all four is the honest mutation. Recording
it because a partial mutation that "survives" reads exactly like a coverage
gap and is not one.

`gltf.rs` was over its ratchet budget with the tests inline, so they moved to
a sibling `gltf_tests.rs` (single `use super::*;`, `geom_color_key` stays
private and is reached through it). No budget raised; ratchet 4/4.

Suite 165 passing. Clippy reports the same 29 pre-existing
`chunks_exact_to_as_chunks` errors as before, none citing the new tests.
